smplayer.svg
Well, I'm not a designer, so I use the vlc.svg and modify it to create the smplayer.svg icon in Inkscape. I use the http://flatuicolors.com/ to chose the colours. I don't know if you will approve it, but i think that looks nice and the spirit of Flatr was respected.