Create smplayer.svg
Well, I'm not a designer, so I create a very simple smplayer.svg icon in Inkscape. I use the http://flatuicolors.com/ to chose the colours. I don't know if you will approve it, but i think that looks nice and the spirit of Flatr was respected.

Sorry about the other closed request, so this is the final version. :)