Create folder-drag-accept.svg
the folder-drag-accept.svg icon is missing. So I create it. When you drag a file to a folder but without dropping the file into the folder, the folder icon should change by creating the illusion that the folder is opened to receive the file. Is this open folder that I leave here.

I hope you approve it.